Logic and Click events
I chnaged the logic for how I am going to connect the GUI and the code.  I changed to make the name of the button follow a different standard and then I parse through the name to determine the row and column value for each cell.  GameDriver class will handle converting column to letters.  I want the columns to be letters so it is easier for the user to see the projected next moves for the AI.  The click event will now change the gui and the array.