Winning Logic Changed
I changed the logic for winning so if a win condition was met the computer would not have to look at the other win conditions.

Added a function in Form1.cs to perform certain tasks if a win condition was met.  I should be able to call this function after the computer makes its move too.

I need to have a draw (cat) functionality.